[SCB-12155] [cmake] Boost to use /dev/[u]random for UUID generation (take 2)
Since boost's UUID library is a header-only built dependency, it's not
enough adding -DBOOST_UUID_RANDOM_PROVIDER_FORCE_POSIX into b2's
cxxflags while building boost in Kudu's thirdparty.  This patch
addresses the issue, defining the BOOST_UUID_RANDOM_PROVIDER_FORCE_POSIX
macro for the preprocessor while building Kudu binaries themselves.
This is similar to the treatment of a similar macro for the boost's
date_time library (the latter is both a binary and a header dependency).

Maybe, the issue might be solved by solely adding the corresponding
flag into the b2's 'define' flag, but I decided to proceed the same
way as it is done for another macro for the boost library.

This is a follow-up to 27418145bc60ceb009c663626a88e57748a85c9f.

Change-Id: I5eedb644bd9f1cd7a48254580bc8307053ce6c9f
Reviewed-on: http://gerrit.cloudera.org:8080/17687
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Alexey Serbin <aserbin@cloudera.com>