WIP [duplication] KUDU-3290 Unite all patches as one to prepare these patches
This patch is for duplication, it includes all about duplication. The purpose
of this patch is do some preparing works for duplication. Some refactors need
to do such as tests.

This patch is a little too many codes, so I plan to split this patch into
10 patches. They are:

1. Adding 'Duplicator' which will write records to kafka,
  adding extra 2 thirdparty libraries: cppkafka, librdkafka.
  librdkfka supports c language apis to operate kafka, and
  cppkafka is a C++ wrapper for librdkakfa, its an api sugger.

2. Add a new raft role DUPLICATOR and raft processing
3. recovering duplication when a leader switches
4. creating/altering in server-side for duplication
5. cpp client apis and CLI tool
6. java client apis
7. suppports kerberos
8. alter a duplication
9. registe kafka uri and validate and unregiste

During this progressing, plans maybe adjust according to advices.

Change-Id: I2ec587821f41eef75078152d7912c11456906ecf