[web] add maintenance op statistics information at web pages for data_retained_bytes
MaintenanceOpStats has collected statistics about 'data_retained_bytes_',
which means the amount of disk capacity can be freed later. But the protobuf
file has no suchinformation, so it is not shown at web pages.

This patch adds a field in the protobuf file and updates the embedded web server
to show the retained bytes for a maintenance operation.

Change-Id: Iac8f0307705d57cea48901102b170c88d73d8c2e